Move code around, make module names shorter
