+++ ariadne/__init__.py
@@ -1,8 +1,6 @@
-from .add_resolve_functions_to_schema import add_resolve_functions_to_schema
-from .build_schema_from_type_definitions import build_schema_from_type_definitions
-from .default_resolver import default_resolver
-from .make_executable_schema import make_executable_schema
-
+from .build_schema import build_schema_from_type_definitions
+from .executable_schema import make_executable_schema
+from .resolvers import add_resolve_functions_to_schema, default_resolver
 
 __all__ = [
     "add_resolve_functions_to_schema",
