+++ ariadne/build_schema.py
@@ -12,6 +12,29 @@ from graphql.language.ast import (
 from graphql.utils.build_ast_schema import build_ast_schema
 
 
+def build_schema_from_type_definitions(
+    type_defs: Union[str, List[str]]
+) -> GraphQLSchema:
+    if isinstance(type_defs, list):
+        type_defs = concatenate_type_defs(type_defs)
+
+    document = parse(type_defs)
+
+    if not document_has_schema(document):
+        schema_definition = build_default_schema(document)
+        document.definitions.append(schema_definition)
+
+    return build_ast_schema(document)
+
+
+def concatenate_type_defs(type_defs: Union[str, List[str]]) -> str:
+    resolved_type_defs = []
+    for type_def in type_defs:
+        if isinstance(type_def, str):
+            resolved_type_defs.append(type_def.strip())
+    return "\n".join(resolved_type_defs)
+
+
 def build_default_schema(document: Document) -> SchemaDefinition:
     defined_types = [
         td.name.value
@@ -38,29 +61,3 @@ def build_default_schema(document: Document) -> SchemaDefinition:
 
 def document_has_schema(document: Document) -> bool:
     return any(isinstance(td, SchemaDefinition) for td in document.definitions)
-
-
-def concatenate_type_defs(type_defs: Union[str, List[str]]) -> str:
-    resolved_type_defs = []
-    for type_def in type_defs:
-        if isinstance(type_def, str):
-            resolved_type_defs.append(type_def.strip())
-    return "\n".join(resolved_type_defs)
-
-
-def build_schema_from_type_definitions(
-    type_defs: Union[str, List[str]]
-) -> GraphQLSchema:
-    if isinstance(type_defs, list):
-        type_defs = concatenate_type_defs(type_defs)
-
-    if isinstance(type_defs, str):
-        document = parse(type_defs)
-    else:
-        document = type_defs
-
-    if not document_has_schema(document):
-        schema_definition = build_default_schema(document)
-        document.definitions.append(schema_definition)
-
-    return build_ast_schema(document)
